// src/app/models/Producto/ProductoVentaGrid.ts
export class ProductoVentaGrid {
  CpInterno: number;
  Codigo: string;
  Detalle: string;
  Unitario: number;
  Neto: number;
  Costo: number;
  Bonificar: number;
  BonificacionMaxima: number;
  StkActual: number;
  StkMinimo: number;
  Iva: number;
  Familia: string;
  IdVariante: number;
  Unidad: string;
}